came up with a stupid jank solution to record videos
basically u run clicker.py in the terminal, move ur cursor to the "start recording" button in obs studio, and watch the magic as the script clicks on the recording to start and stop it.  then, copy paste trim_video.py into the folder with all the videos and it runs ffmpeg to trim all to exactly 90 frames @ 30fps = 3 second video.  it's a stupid jank solution but it works so stfu